Add initial api version
This commit adds the initial api and route to return all data.

The long term plan isn't to return the whole of the data at once,
but to add routes with route or query parameters for each table or
column.

That of course depends on freetime and other projects. This and
acnh-data-tools are just a small diversion while my SO plays
ACNH :)