Add all routes and table lookup name
Rather than add a route file for each route to be mounted as the
route name, this commit uses a route parameter to the root route.

Since the route isn't the same as the table name, this commit also
adds a function to look up the table name for the route.

If nothing is returned from the lookup, the route is unknown and
the user gets a 404 NOT FOUND.